+++ app/src/main/java/com/mindorks/framework/mvp/ui/manager/restaurant/ManagerRestaurantMvpView.java
@@ -5,7 +5,7 @@ import com.mindorks.framework.mvp.ui.base.MvpView;
 
 public interface ManagerRestaurantMvpView extends MvpView {
 
-    public void openPromotionDetailsActivity(int promotionId);
+    public void openPromotionDetailsActivity(Long promotionId);
     public void openDailyMenuDetailsActivity(int dailyMenuId);
 
 }
